All token's actions are available from the token detail (preview). This commit closes #63.
